Add extension support and 'request' extension.
Extensions are optional parts of the library. This mechanism was added
to put common code, such as loading JSON-LD, into this library.

Currently only the 'request' module is available when using node.js.
This extension can load JSON-LD from stdin ('-'), JSON-LD from files,
URLs with JSON-LD, or URLs with RDFa and and will return parsed raw
JSON-LD.

The basic usage for the 'request' extension is:

    var jsonld = require('jsonld');
    jsonld.use('request');
    jsonld.request('...', function(err, data) {
      ...
    }